+++ test/recurly/transaction_test.php
@@ -16,18 +16,38 @@ class Recurly_TransactionTest extends UnitTestCase
     $this->assertEqual($transaction->account->getHref(), 'https://api.recurly.com/v2/accounts/verena');
   }
   
-  public function testFailedTransaction()
+  public function testCreateTransactionFailed()
   {
     $responseFixture = loadFixture('./fixtures/transactions/create-422.xml');
 
     $client = new MockRecurly_Client();
+    $client->returns('request', $responseFixture);
+
+    $transaction = new Recurly_Transaction(null, $client);
+
+    try {
+      $transaction->create();
+      $this->fail("Expected Recurly_ValidationError");
+    }
+    catch (Recurly_ValidationError $e) {
+      $this->pass("Received Recurly_ValidationError");
+      $this->assertEqual($e->getMessage(), 'Your card number is not valid. Please update your card number.');
+      $this->assertIsA($e->errors[0], 'Recurly_TransactionError');
+      $this->assertIsA($e->object, 'Recurly_Transaction');
+    }
+  }
+  
+  public function testGetFailedTransaction()
+  {
+    // GET response is "200 OK", yet transaction had an error
+    $responseFixture = loadFixture('./fixtures/transactions/show-200-error.xml');
+
+    $client = new MockRecurly_Client();
     $client->returns('request', $responseFixture, array('GET', '/transactions/012345678901234567890123456789ab'));
 
-    $transaction_errors = Recurly_Transaction::get('012345678901234567890123456789ab', $client);
+    $transaction = Recurly_Transaction::get('012345678901234567890123456789ab', $client);
 
-    $this->assertIsA($transaction_errors, 'Recurly_ErrorList');
-    $this->assertIsA($transaction_errors->transaction, 'Recurly_Transaction');
-    $this->assertIsA($transaction_errors->transaction_error, 'Recurly_TransactionError');
-    $this->assertEqual($transaction_errors->transaction->getHref(), 'https://api.recurly.com/v2/transactions/abcdef1234567890');
+    $this->assertIsA($transaction, 'Recurly_Transaction');
+    $this->assertIsA($transaction->transaction_error, 'Recurly_TransactionError');
   }
 }